requirements.txt
streamlit
google-generativeai
pandas
gTTS
SpeechRecognition
streamlit-mic-recorder